requirements.txt
discord.py
ollama
python-dotenv
